ISSUE #3627 - Extract newMenu component
