+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsList/newTicketMenu/newTicketMenu.component.tsx
@@ -1,0 +1,52 @@
+/**
+ *  Copyright (C) 2022 3D Repo Ltd
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { TicketsHooksSelectors } from '@/v5/services/selectorsHooks/ticketsSelectors.hooks';
+import { useContext } from 'react';
+import { FormattedMessage } from 'react-intl';
+import { useParams } from 'react-router-dom';
+import { CardContext } from '@components/viewer/cards/cardContext.component';
+import AddIcon from '@assets/icons/add_circle.svg';
+import { sortBy } from 'lodash';
+import { ActionMenu, MenuItem, NewTicketButton } from '../ticketsList.styles';
+import { TicketsCardViews } from '../../tickets.constants';
+import { ViewerParams } from '../../../../routes.constants';
+
+export const NewTicketMenu = () => {
+	const contextValue = useContext(CardContext);
+	const { containerOrFederation } = useParams<ViewerParams>();
+	const templates = TicketsHooksSelectors.selectTemplates(containerOrFederation);
+
+	const goToNewTicket = (template) => contextValue.setCardView(TicketsCardViews.New, { template });
+
+	return (
+		<ActionMenu
+			TriggerButton={(
+				<NewTicketButton>
+					<AddIcon />
+					<FormattedMessage id="viewer.cards.tickets.newTicket" defaultMessage="New Ticket" />
+				</NewTicketButton>
+			)}
+		>
+			{sortBy(templates, 'name').map((template) => (
+				<MenuItem onClick={() => goToNewTicket(template)} key={template._id}>
+					{template.name}
+				</MenuItem>
+			))}
+		</ActionMenu>
+	);
+};
