+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsList/ticketsListCard.component.tsx
@@ -19,44 +19,15 @@ import { TicketsActionsDispatchers } from '@/v5/services/actionsDispatchers/tick
 import { TicketsHooksSelectors } from '@/v5/services/selectorsHooks/ticketsSelectors.hooks';
 import { modelIsFederation } from '@/v5/store/tickets/tickets.helpers';
 import { CardContainer, CardHeader } from '@components/viewer/cards/card.styles';
-import { useContext, useEffect } from 'react';
+import { useEffect } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { useParams } from 'react-router-dom';
 import TicketsIcon from '@mui/icons-material/FormatListBulleted';
 import { CardContent } from '@components/viewer/cards/cardContent.component';
 import { UsersActionsDispatchers } from '@/v5/services/actionsDispatchers/usersAction.dispatchers';
-import { CardContext } from '@components/viewer/cards/cardContext.component';
-import AddIcon from '@assets/icons/add_circle.svg';
-import { sortBy } from 'lodash';
 import { TicketsList } from './ticketsList.component';
-import { ActionMenu, MenuItem, NewTicketButton } from './ticketsList.styles';
-import { TicketsCardViews } from '../tickets.constants';
+import { NewTicketMenu } from './newTicketMenu/newTicketMenu.component';
 import { ViewerParams } from '../../../routes.constants';
-
-const NewTicketMenu = () => {
-	const contextValue = useContext(CardContext);
-	const { containerOrFederation } = useParams<ViewerParams>();
-	const templates = TicketsHooksSelectors.selectTemplates(containerOrFederation);
-
-	const goToNewTicket = (template) => contextValue.setCardView(TicketsCardViews.New, { template });
-
-	return (
-		<ActionMenu
-			TriggerButton={(
-				<NewTicketButton>
-					<AddIcon />
-					<FormattedMessage id="viewer.cards.tickets.newTicket" defaultMessage="New Ticket" />
-				</NewTicketButton>
-			)}
-		>
-			{sortBy(templates, 'name').map((template) => (
-				<MenuItem onClick={() => goToNewTicket(template)} key={template._id}>
-					{template.name}
-				</MenuItem>
-			))}
-		</ActionMenu>
-	);
-};
 
 export const TicketsListCard = () => {
 	const { teamspace, project, containerOrFederation } = useParams<ViewerParams>();
